requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
python-dotenv>=1.0

# Oracle DB — uncomment when using Oracle
# cx_Oracle>=8.3
# or newer: oracledb>=1.4

# Google OAuth — uncomment if using Google login
# google-auth>=2.23
